feat: auth wrapper
